# Dramatically increase re-sync performance