--- a/CHANGES/6373.bugfix
+++ b/CHANGES/6373.bugfix
@@ -1,0 +1,1 @@
+Improved resync performance by up to 2x with a change to the content stages.