@dralley 2020-03-19
## Performance tests

TL;DR improves the speed of file plugin re-sync nearly double, RPM plugin re-sync and re-migration by around 33%

### File Plugin, 30k units, on_demand

Before:

- Initial sync: 72 seconds
- Re-sync: 63 seconds

After:

- Initial sync: 72 seconds
- Re-sync: 33 seconds

### Rpm Plugin, ~10k units (centos), on_demand

Before:

- Initial sync: 61 seconds
- Re-sync: 48 seconds

After:

- Initial sync: 53 seconds
- Re-sync: 33 seconds

### Migration plugin, migration of 1 repository, ~10k units (centos) from Pulp 2

Before:

- Initial migration: 88 seconds
- Re-migration: 77 seconds

After:

- Initial migration: 88 seconds
- Re-migration: 54 seconds