@dralley 2020-03-19
This probably deserves a changelog entry but there's no issue filed.  Should I file an issue? Is there a way to add an entry without doing so?