@bmbouter 2020-03-19
I think performance fixes would benefit from a changelog entry. We use `[noissue]` less and less these days. The code changes look great tho!