@bmbouter 2020-03-19
Is it correct that this whole workload runs in a single worker? If so, in terms of speeding this up significantly, my recommendation would be to explore running multiple pipelines in parallel across several workers each handling disjoint parts of the pulp2 migration.