@dralley 2020-03-19
@bmbouter That's certainly something to explore but it would be great to exhaust all the opportunities present here first, since every improvement made here will also speed up re-sync performance by a huge margin.