@bmbouter 2020-03-19
Improving this will help all workflows I agree. I observe though that with each change it's improving by only a few percent. What that tells me is that we'll have to start to do draconian things to our code, e.g. postgres stored procedures, etc, etc to get improvements beyond a few percent. If you want to continue down that road that's ok w/ me. The usual techniques to use there are a mixture of cprofile for improve the python code's runtime, and then SQL EXPLAIN ANALYZE on the queries executed in the stages taking the most time.

My take on the workload is that it is heavily db constrained, and yet we have lots of indexes already and I don't see one missing. Adding indexes is the most common way I know how to make a query "go faster", but I don't see that opportunity since we already have indexes on all the tables involved: https://gist.github.com/bmbouter/20e7743efb0906f25cbd4c45860cfd82

So how to you solve a large problem faster? Divide it into sub-problems! Also from a human effort and benefit to our users perspective a parallel worker approach would give a much higher return I think.