@bmbouter 2020-03-20
@dalley I would lgtm this if it could have a changelog entry. How do you feel about adding that?