@fabricio-aguiar 2020-03-20
> Even with this improvement, the QueryExistingContents() stage is really expensive. It's 50% (_exactly_ 50%) of the time spent in a resync (of the file plugin), and 33% of the time used in a re-migration (of an RPM repository). @bmbouter @fabricio-aguiar do you have any ideas on how to speed it up further?
> 
> Resync:
> ![file_resync](https://user-images.githubusercontent.com/2751996/77024523-2add9700-6965-11ea-9b97-9ecfc72dd39b.png)

The only thing I can think is: replace `OR` with `IN` on:
```python
content_q_by_type[model_type] = content_q_by_type[model_type] | unit_q
```